#6547-Add-Immunization-Test-Fix : created area type enum, updated 2 duplicated enums (laboratory), updated person's area type to be dynamic, created dynamic gps coordonates generation
